feat(agenda): add in conversion of org-tag-faces
Sending the tag data to the client is the only viable option when trying
to get face data over. Because `org-tag-faces` allows for existing
faces, custom single colors, and custom multiple attributes, it's
impossible to get the foreground color of a tag without referencing the
variable. If you're referencing the variable, you might as well just
send it over and let the client deal with it all.

Reasonable chance that the tag-faces may not cover every case. In
particular, it might be possible to set tag faces using odd syntax that
this won't account for.