feat(tsx): add in client display of tag color
Makes a few assumptions about the shape of tags coming in from the
server but should work with most tag inheritance scenarios.

If no matching tag color is found, should go back to the CSS default.

Properly splits out data and item type when sending messages from the
server to the client. Should get this more strongly typed than currently
but that's a low priority for the current feature set.