feat(tsx): add in local storage for matchQuery
This just sets the matchQuery option but will hit the websocket in the
next commit. The value persists in local rather than sync storage as the
user may want to set different values depending on the machine.

Added a ref for the input as the defaultValue is not the stored value
on first render. This should hopefully avoid any errors for the meantime
as the input form value should only be set after the user explicitly
sets it (so a mostly uncontrolled input).