feat(ws): add ports connection to background script
Using the Ports API to talk to the background script. Should only send
and listen once with the onInitialRender ref.

Background script will server as handler between master/slave tabs with
the idea being that only one tab has a websocket connection and the
others will talk to it. This should kill the emacs bugs wherein multiple
tabs each open their own websocket and try to talk to Emacs which thinks
there is only one socket.