feat(background): buggy implementation of chat
This is a halfway point as the Port API is extremely buggy and it's
probably better to try and get a content script working with the new tab
page. Doing so will allow storing tab id in storage and will heavily
reduce the need for the background script to persist for long periods.