refactor(background): massive simplification of query flow
Most bugs starting to be sorted. Worst one was multiple background
scripts being created by newtab page importing from background script.

Currently, the newtab uses ports to talk to the background. The
background uses messaging to talk to the newtabs. When adding back
storage, this should make it possible to maintain a list of ids even
if the background script is reloaded. Lot more work here to do.

The Promise on the background doesn't race yet. Unsure if it should be
added back in. It's not clear if it's needed.