feat(background): full master->client handoff
This should now be complete -- allows master to reload with other tabs
open and the master to be re-identified. This could be made more
efficient by storing the master but it's not clear that's necessary.
State can live in the new tab side which is persistent unlike the
background worker.

Will need to handle the case of master closing and a new one being
picked when that happens.