refactor(wscontext): remove switching context
Currently in an unfinished state/regression. This should stop the
remounting of index. The idea is that the data now needs to live in
the wscontext and the switching should be invisible to the consumers.

React has no way to allow moving a component between different trees
without resorting to unstable APIs.