refactor(usesinglewebsocket): switch to hooks
Switching between two contexts makes everything remount. This refactors
the relevant code into a hook.

useWebSocket is called with null if it's a client which isn't a great
solution but will have to do for now. It doesn't throw up errors and
seems to switch when coming back to master.