feat(storage): sync state between tabs
Using chrome local storage as a way to sync state between the tabs
rather than using message passing with the background worker due to the
possible flakiness of the background worker. Normally getting from the
local storage is an async operation but library has abstracted the
difficulties of it out. Lots of write operations should be avoided as
they are blocking but that shouldn't be a problem here.

Next step is pushing the state up to a master context and allowing
components to use directly from it. Once this is completed, a lot of
simplification should be possible with the message passing code on both
the front and back end.