fix(optionsmenu): pass updatematchquery as prop
Previously, the state update of matchquery was being done asynchronously
and its un-definite value was being used to send data to emacs. This was
causing all sorts of race conditions and bugs. By passing in the update
function as an 'action' and forcing the form data as the argument, this
clears a whole class of bugs.

Should be mindful of other bugs in the same general area -- relying on
storage to sync the tabs has this as a major downside.