feat(newtab): add in tab to tab messages
This allows a new match query to be sent from the client to the master
websocket tab. The master websocket tab then handles the sending of data
to Emacs as normal.

Required fixing some tricky bugs around the settings updating process.
It's crucial not to screw up updating the settings -- it must be done
only once as it'll cause massive amounts of re-renders otherwise.