feat(store): add redux
The idea is to eventually replace the store as context with Redux. It's
clear that the current architecture will break as the app grows.

Have replaced the value in OptionsMenu, will type more and break out
reducers into multiple slices as needed. For now, getting the storage
syncing working is a priority.

Didn't originally go with this solution due to potential issues syncing
redux state with storage. Apparently there are some annoying bugs in the
libs involved but they likely won't apply to this application. If the
storage resets between a load, it's not the end of the world for the
application as it stands now. It may become so but hopefully moving off
of redux will be straightforward when that need arises.