feat(middleware): add middleware to handle socket events
Moved from react-use-websocket to handling websocket connections
manually in Redux middleware. This is more akin to the 'Redux' way of
doing things.

Will require quite a bit of rework but should eventually means minimal
to no use of contexts for websocket related issues.

It's probably optimal to move the background negotiation to either
a hook or middleware as well.

Next steps will be building out the websocket behavior for error
handling and reconnection.