refactor(app): break out slices into modules
The emacs slice is more 'emacs data' as the WS which deals with Emacs
is a separate concern. However, the WS slice may still need to be broken
up. Likely there is room for more refactoring.

Assuming a hybrid module approach where view data isn't lumped together
with the features unlike what RTK recommends. Given this is an unusual
usage of RTK, it should be a better path.

Likely won't use RTK query either due to on/off role of the
master/client WS. This middleware approach does what is needed,
possibly with too many side effects but I'd rather fix that than try
to shoehorn this into RTK query.