fix(rootreducer): allow deeper persist merge
For some of the entries like orgitem and tags, the state can be several
levels deep so use the mergelevel2 import to stop clobbering through.

Can be resolved in the future using more specific state for things like
title but the tags object will likely always need this.

Types were specified for the persistence reducer to keep TS happy.