feat(optionsmenu): break out into multiple menus
Working off a Figma design and slowly implementing it. This is in an
unfinished state and requires reworking the tests.

Breaking the content and the bar into 2 will allow for using pure CSS
for animations. The downside is added complexity managing state of menu.
Thinking of using Redux for this but going to keep playing around with
it for now.

Still a long way from implementing layout and theming options so those
buttons are going to render null for now.