feat(layout): implement dnd-kit
Requires a lot more work to get position and size right. General idea
is that the component has 3 areas to go to and in the future, can be
ordered within those areas. Trying to get connection status working
with this first.

Potential problem is storage is required to persist these settings so
will need to figure out ways to work around flashes of unorganized
items before the storage is loaded.