feat(debug): add clear storage button
Added extraReducers to all slices to allow for a single action that
resets all of them. Shouldn't be too much extra cognitive load when
creating new slices and it's likely the most correct way to do it for
Redux toolkit aside from possibly using a middleware.

Button will clear storage using a listener instead of redux-persist's
purge method since it's more straightforward in this case and context.