feat(el/server): add async queue
This is not a mutex/lock/semaphore. It's more akin to a queue but the
gist is that the current async task starts and an id for it is saved.

If a new async task comes in, it overrides the old task. Cancelling the
old task isn't possible but if the old task discovers it's not longer
the priority, it will simply do nothing and drop the result.

When a synchronous clocked in task comes in, it'll automatically
override the old task by setting the queue to nil. This allows for the
case where a user is in a clocked in task, then clocks into another task
directly. This will call the clock out hook (async) and the clock in hook
(sync) after it but the clock in hook will now take precedence as the
user expects.

Given how complicated this could've been, this seems elegant if a bit
hacky.