test(e2e): add full emacs to chrome e2e test
Currently relies on the test to kill the emacs process and has no way
of dealing with errors on the emacs side but the test shouldn't pass
if there are errors as the front end shouldn't get the right data.

Needs port randomization in the future as well as some refactoring
around the process generation.

Overall though, finally a MUCH needed full integration test.