fix(ws): allow for dropped responses
This is a hack to allow fast changes in match query or slow responses
from Emacs from the client to expect responses that were dropped on the
Emacs side.

Added a todo that in the future, a more built out responses tracker
is needed that can track responses based on the query (ie drop repeated
queries for match queries but don't drop repeated queries for other
data).