fix(ws): prevent race conditions when clocking in
If a getItem is sent from the client to Emacs and the user clocks in
before Emacs responds, the original request is dropped. This is
intentional but the client doesn't know about a dropped request so
this is a hack to basically drop expected responses on clocked in data.

This will have to be redone when the response code is rewritten but
it resolves a lot of test flakiness for now.