fix(rootreducer): stick to single level merge
When originally adding in two level merge in
5058c87088985ad20c1a18dae1182f5cb7d3acac , the emacs state had just
saved the entire org item object and a very nested tagFaces object.

Current implementation attempts to flatten the org item and tagFaces
should be overriden when new data comes in.

Was finding a bug switching between different running instances very
quickly where the tag items object would be merged which would then
cause a rehydration loop.

Given the usage of redux-persist, if there is some reason why it should
be added back in, engineer around it and attempt to use a flattened
state if at all possible.