test(e2e): slow down tab test for new playwright ver
Playwright v1.40+ has slowed down tab management by quite a bit in
Webkit and Chromium. This increases the test timeout and seems to make
it happy.

The exact issue is more complicated -- using DWM's monocle view it works
without having to increase the timeout but in the tiled view, it needs
the extra timeout. There are several issues somewhat related to this
in the Playwright repo so we'll wait for them to deal with it.