fix(ws): force storage flush on new emacs msg
Because of the slightly flaky nature of redux-persist, a rehydration
action could occur before new data from Emacs was saved. This would only
be seen in testing-environment usage due to the speed required to
trigger the bug.

With this application, there is a relatively easy fix: trigger a flush
(write to storage) action when new data comes in from Emacs. That's the
only point where stored data can be inadvertently overridden like this.

Redux-persist's native FLUSH action isn't usable outside of the library
so a custom action using persistor.flush() in middleware is created.
This requires a relatively dodgy method of passing the persistor value
since it's based upon the store that already has middleware initialized
but using a class should be an ok solution for now.

Also added logging middleware for easier debugging of this sort of
problem in the future.