fix(persistor): pause all rehydration upon flush
Under certain circumstances, the rehydration process could be triggered
after flushing had started but before storage was actually written to.
This pauses that process in the middleware and Persistor class layer
until the flush is complete.

The need for queueing rehydration isn't needed as calling .persist will
trigger a rehydration (which is the same as resync in the plasmo fork).