refactor(lisp/store): create store
As the code gets more complex, some way to manage the hooks from org and
the requests from the extension is needed. Caching in particular will
benefit here as it's slowly implemented.

This follows the Flux/Redux architecture already in use on the frontend.
It could probably use with some tweaking since throwing around action
objects makes less sense here but it works to the same effect thus far.