feat(ws): add in finding type for async tasks
When an async find for a new item has been triggered on the Emacs side
(from a hook, say clock out), this lets the front end know that an item
is coming (thus it should show the loading bar).

On the Emacs side, could use more dispatches to do this but this is
straightforward so no need to overcomplicate it.

On the client side, the checking of type in the wsSlice is inelegant
but it'll do for now. As more functionality is built up, this can be
changed.